+++ modules/compute_disk_snapshot/versions.tf
@@ -17,13 +17,10 @@
 terraform {
   required_version = ">=0.13.0"
   required_providers {
-    google      = ">= 3.43, <4.0"
-    google-beta = ">= 3.43, <4.0"
+    google = ">= 3.43, <4.0"
+    null   = ">= 2.1"
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-vm:preemptible_and_regular_instance_templates/v6.2.0"
-  }
-  provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-vm:preemptible_and_regular_instance_templates/v6.2.0"
+    module_name = "blueprints/terraform/terraform-google-vm:compute_instance/v6.2.0"
   }
 }