+++ modules/compute_instance/outputs.tf
@@ -19,6 +19,11 @@ output "instances_self_links" {
   value       = google_compute_instance_from_template.compute_instance.*.self_link
 }
 
+output "instances_details" {
+  description = "List of all details for compute instances"
+  value       = google_compute_instance_from_template.compute_instance.*
+}
+
 output "available_zones" {
   description = "List of available zones in region"
   value       = data.google_compute_zones.available.names