+++ modules/mig/versions.tf
@@ -21,9 +21,9 @@ terraform {
     google-beta = ">= 3.43, <4.0"
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-vm:mig/v6.0.0"
+    module_name = "blueprints/terraform/terraform-google-vm:mig/v6.2.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-vm:mig/v6.0.0"
+    module_name = "blueprints/terraform/terraform-google-vm:mig/v6.2.0"
   }
 }