+++ test/fixtures/compute_instance/disk_snapshot/main.tf
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">=0.13.0"
-  required_providers {
-    google      = ">= 3.43, <4.0"
-    google-beta = ">= 3.43, <4.0"
-  }
-  provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-vm:preemptible_and_regular_instance_templates/v6.2.0"
-  }
-  provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-vm:preemptible_and_regular_instance_templates/v6.2.0"
-  }
+
+module "disk_snapshot" {
+  source     = "../../../../examples/compute_instance/disk_snapshot"
+  project_id = var.project_id
+  region     = "us-central1"
+  subnetwork = google_compute_subnetwork.main.self_link
 }